Extract double-coding review API service
