+++ apps/frontend/src/app/coding/services/test-person-coding.service.spec.ts
@@ -877,199 +877,6 @@ describe('TestPersonCodingService', () => {
     });
   });
 
-  describe('getDoubleCodedVariablesForReview', () => {
-    it('should request double-coded review data with agreement and scope filters', () => {
-      const mockResponse = {
-        data: [
-          {
-            responseId: 10,
-            unitName: 'UNIT_1',
-            variableId: 'VAR_1',
-            personLogin: 'person-1',
-            personCode: 'P001',
-            bookletName: 'BOOKLET_1',
-            givenAnswer: 'answer',
-            isResolved: false,
-            appliedCode: null,
-            appliedScore: null,
-            appliedComment: null,
-            coderResults: [
-              {
-                coderId: 1,
-                coderName: 'Coder 1',
-                jobId: 100,
-                jobName: 'Job A',
-                code: 1,
-                score: 1,
-                notes: null,
-                supervisorComment: null,
-                codedAt: '2026-05-18T00:00:00.000Z'
-              }
-            ]
-          }
-        ],
-        total: 1,
-        page: 2,
-        limit: 25
-      };
-
-      service
-        .getDoubleCodedVariablesForReview(
-          mockWorkspaceId,
-          2,
-          25,
-          true,
-          false,
-          'VAR_1',
-          9,
-          'done',
-          'unresolved',
-          'differ',
-          [11, 12],
-          [21]
-        )
-        .subscribe(response => {
-          expect(response).toEqual(mockResponse);
-        });
-
-      const req = httpMock.expectOne(
-        request => request.url ===
-          `${mockServerUrl}admin/workspace/${mockWorkspaceId}/coding/double-coded-review`
-      );
-      expect(req.request.method).toBe('GET');
-      expect(req.request.params.get('page')).toBe('2');
-      expect(req.request.params.get('limit')).toBe('25');
-      expect(req.request.params.get('onlyConflicts')).toBe('true');
-      expect(req.request.params.get('excludeTrainings')).toBe('false');
-      expect(req.request.params.get('search')).toBe('VAR_1');
-      expect(req.request.params.get('coderId')).toBe('9');
-      expect(req.request.params.get('statusFilter')).toBe('done');
-      expect(req.request.params.get('resolvedFilter')).toBe('unresolved');
-      expect(req.request.params.get('agreementFilter')).toBe('differ');
-      expect(req.request.params.get('jobDefinitionIds')).toBe('11,12');
-      expect(req.request.params.get('coderTrainingIds')).toBe('21');
-      req.flush(mockResponse);
-    });
-
-    it('should propagate double-coded review errors to the component', done => {
-      service.getDoubleCodedVariablesForReview(mockWorkspaceId).subscribe({
-        next: () => done.fail('expected double-coded review request to fail'),
-        error: error => {
-          expect(error.status).toBe(500);
-          done();
-        }
-      });
-
-      const req = httpMock.expectOne(
-        `${mockServerUrl}admin/workspace/${mockWorkspaceId}/coding/double-coded-review?page=1&limit=50&onlyConflicts=false&excludeTrainings=false`
-      );
-      req.flush(
-        { message: 'review query failed' },
-        { status: 500, statusText: 'Server Error' }
-      );
-    });
-  });
-
-  describe('double-coded review decisions', () => {
-    it('should send the authoritative source unit when saving a draft', () => {
-      const draft = {
-        sourceUnitId: 77,
-        code: 3,
-        score: 2,
-        comment: 'Check this'
-      };
-
-      service
-        .saveDoubleCodedReviewDraft(mockWorkspaceId, 10, draft)
-        .subscribe();
-
-      const req = httpMock.expectOne(
-        `${mockServerUrl}admin/workspace/${mockWorkspaceId}/coding/double-coded-review/10/draft`
-      );
-      expect(req.request.method).toBe('PUT');
-      expect(req.request.body).toEqual(draft);
-      req.flush({
-        id: 1,
-        responseId: 10,
-        managerUserId: 99,
-        managerName: 'Reviewer',
-        state: 'draft',
-        code: 3,
-        selectedCode: 3,
-        score: 2,
-        comment: 'Check this',
-        createdAt: null,
-        updatedAt: null,
-        finalizedAt: null,
-        legacy: false
-      });
-    });
-
-    it('should delete the current manager draft', () => {
-      service.deleteDoubleCodedReviewDraft(mockWorkspaceId, 10).subscribe();
-
-      const req = httpMock.expectOne(
-        `${mockServerUrl}admin/workspace/${mockWorkspaceId}/coding/double-coded-review/10/draft`
-      );
-      expect(req.request.method).toBe('DELETE');
-      req.flush({ success: true });
-    });
-
-    it('should post explicit replay code decisions for double-coded review resolutions', () => {
-      const mockResponse = {
-        success: true,
-        appliedCount: 1,
-        failedCount: 0,
-        skippedCount: 0,
-        message: 'ok'
-      };
-      const body = {
-        decisions: [
-          {
-            responseId: 10,
-            sourceUnitId: 77,
-            code: 3,
-            score: 2,
-            resolutionComment: 'Replay checked'
-          }
-        ]
-      };
-
-      service
-        .applyDoubleCodedResolutions(mockWorkspaceId, body)
-        .subscribe(response => {
-          expect(response).toEqual(mockResponse);
-        });
-
-      const req = httpMock.expectOne(
-        `${mockServerUrl}admin/workspace/${mockWorkspaceId}/coding/double-coded-review/apply-resolutions`
-      );
-      expect(req.request.method).toBe('POST');
-      expect(req.request.body).toEqual(body);
-      req.flush(mockResponse);
-    });
-
-    it('propagates transport errors to the caller', () => {
-      const errorSpy = jest.fn();
-
-      service
-        .applyDoubleCodedResolutions(mockWorkspaceId, { decisions: [] })
-        .subscribe({
-          error: errorSpy
-        });
-
-      const req = httpMock.expectOne(
-        `${mockServerUrl}admin/workspace/${mockWorkspaceId}/coding/double-coded-review/apply-resolutions`
-      );
-      req.flush('Unavailable', {
-        status: 503,
-        statusText: 'Service Unavailable'
-      });
-
-      expect(errorSpy).toHaveBeenCalled();
-    });
-  });
-
   describe('getCohensKappaStatistics', () => {
     it('should request detailed kappa statistics with filters and return variable mean kappa', () => {
       const mockResponse = {
