+++ apps/frontend/src/app/coding/services/test-person-coding.service.ts
@@ -43,13 +43,6 @@ import {
 } from '../../../../../../api-dto/coding/coding-freshness.dto';
 import { AutocodingReadinessDto } from '../../../../../../api-dto/coding/autocoding-readiness.dto';
 import {
-  DoubleCodedManagerDecisionDto,
-  DoubleCodedResolutionDecisionDto,
-  DoubleCodedResolutionResultDto,
-  DoubleCodedReviewCodeDto,
-  SaveDoubleCodedReviewDraftDto
-} from '../../../../../../api-dto/coding/double-coded-review.dto';
-import {
   ResponseMatchingFlag,
   WorkspaceSettingsService
 } from '../../ws-admin/services/workspace-settings.service';
@@ -71,8 +64,6 @@ export interface CodingStatistics {
     [key: string]: number;
   };
 }
-
-export type DoubleCodedResolutionDecision = DoubleCodedResolutionDecisionDto;
 
 export interface CodingStatisticsWithJob extends CodingStatistics {
   jobId?: string;
@@ -1698,173 +1689,6 @@ export class TestPersonCodingService {
         })
         )
       );
-  }
-
-  getDoubleCodedVariablesForReview(
-    workspaceId: number,
-    page: number = 1,
-    limit: number = 50,
-    onlyConflicts: boolean = false,
-    excludeTrainings: boolean = false,
-    search?: string,
-    coderId?: number,
-    statusFilter?: string,
-    resolvedFilter?: string,
-    agreementFilter?: 'all' | 'match' | 'differ',
-    jobDefinitionIds?: number[],
-    coderTrainingIds?: number[]
-  ): Observable<{
-      data: Array<{
-        responseId: number;
-        sourceUnitId: number;
-        unitName: string;
-        variableId: string;
-        personLogin: string;
-        personCode: string;
-        bookletName: string;
-        givenAnswer: string;
-        isResolved: boolean;
-        appliedCode: number | null;
-        appliedScore: number | null;
-        appliedComment: string | null;
-        availableCodes: DoubleCodedReviewCodeDto[];
-        managerDrafts: DoubleCodedManagerDecisionDto[];
-        managerHistory: DoubleCodedManagerDecisionDto[];
-        coderResults: Array<{
-          coderId: number;
-          coderName: string;
-          jobId: number;
-          jobName: string;
-          code: number | null;
-          codingIssueOption: number | null;
-          score: number | null;
-          notes: string | null;
-          supervisorComment: string | null;
-          codedAt: string;
-        }>;
-      }>;
-      total: number;
-      page: number;
-      limit: number;
-    }> {
-    let params = new HttpParams()
-      .set('page', page.toString())
-      .set('limit', limit.toString())
-      .set('onlyConflicts', onlyConflicts.toString())
-      .set('excludeTrainings', excludeTrainings.toString());
-
-    if (search && search.trim() !== '') {
-      params = params.set('search', search.trim());
-    }
-
-    if (coderId) {
-      params = params.set('coderId', coderId.toString());
-    }
-
-    if (statusFilter && statusFilter !== 'all') {
-      params = params.set('statusFilter', statusFilter);
-    }
-
-    if (resolvedFilter && resolvedFilter !== 'all') {
-      params = params.set('resolvedFilter', resolvedFilter);
-    }
-
-    if (agreementFilter && agreementFilter !== 'all') {
-      params = params.set('agreementFilter', agreementFilter);
-    }
-
-    if (jobDefinitionIds?.length) {
-      params = params.set('jobDefinitionIds', jobDefinitionIds.join(','));
-    }
-
-    if (coderTrainingIds?.length) {
-      params = params.set('coderTrainingIds', coderTrainingIds.join(','));
-    }
-
-    return this.http.get<{
-      data: Array<{
-        responseId: number;
-        sourceUnitId: number;
-        unitName: string;
-        variableId: string;
-        personLogin: string;
-        personCode: string;
-        bookletName: string;
-        givenAnswer: string;
-        isResolved: boolean;
-        appliedCode: number | null;
-        appliedScore: number | null;
-        appliedComment: string | null;
-        availableCodes: DoubleCodedReviewCodeDto[];
-        managerDrafts: DoubleCodedManagerDecisionDto[];
-        managerHistory: DoubleCodedManagerDecisionDto[];
-        coderResults: Array<{
-          coderId: number;
-          coderName: string;
-          jobId: number;
-          jobName: string;
-          code: number | null;
-          codingIssueOption: number | null;
-          score: number | null;
-          notes: string | null;
-          supervisorComment: string | null;
-          codedAt: string;
-        }>;
-      }>;
-      total: number;
-      page: number;
-      limit: number;
-    }>(
-      `${this.serverUrl}admin/workspace/${workspaceId}/coding/double-coded-review`,
-      { headers: this.authHeader, params }
-    );
-  }
-
-  applyDoubleCodedResolutions(
-    workspaceId: number,
-    dto: { decisions: DoubleCodedResolutionDecision[] }
-  ): Observable<{
-      success: boolean;
-      appliedCount: number;
-      failedCount: number;
-      skippedCount: number;
-      message: string;
-      results: DoubleCodedResolutionResultDto[];
-    }> {
-    return this.http.post<{
-      success: boolean;
-      appliedCount: number;
-      failedCount: number;
-      skippedCount: number;
-      message: string;
-      results: DoubleCodedResolutionResultDto[];
-    }>(
-      `${this.serverUrl}admin/workspace/${workspaceId}/coding/double-coded-review/apply-resolutions`,
-      dto,
-      { headers: this.authHeader }
-    );
-  }
-
-  saveDoubleCodedReviewDraft(
-    workspaceId: number,
-    responseId: number,
-    draft: SaveDoubleCodedReviewDraftDto
-  ): Observable<DoubleCodedManagerDecisionDto> {
-    return this.http.put<DoubleCodedManagerDecisionDto>(
-      `${this.serverUrl}admin/workspace/${workspaceId}/coding/double-coded-review/${responseId}/draft`,
-      draft,
-      { headers: this.authHeader }
-    );
-  }
-
-  deleteDoubleCodedReviewDraft(
-    workspaceId: number,
-    responseId: number
-  ): Observable<{ success: boolean }> {
-    return this.http.delete<{ success: boolean }>(
-      `${this.serverUrl}admin/workspace/${workspaceId}/coding/double-coded-review/${responseId}/draft`,
-      { headers: this.authHeader }
-    );
   }
 
   getCohensKappaStatistics(
